Add Schedule Tab
Add a tab for schedule that lists off all scheduled jobs, when they are
scheduled to run, and on which queue.

It reuses the same job modal and adds a "Perform at" item if it's a scheduled
job and removes the "inspect" button since that won't really provide more
context than is already visible for a scheduled job.